发送视频区分类型

+++ leosupport/src/main/java/leo/work/support/biz/MediaProjectionService.java
@@ -77,9 +77,9 @@ public class MediaProjectionService extends Service {
     }
 
     // 媒体类型 ---- 视频
-    public static int TYPE_VIDEO = 0;
+    public static byte TYPE_VIDEO = 0;
     // 媒体类型 ---- 音频
-    public static int TYPE_AUDIO = 1;
+    public static byte TYPE_AUDIO = 1;
     // 录频实例
     private MediaProjectionManager manager;
     // 数据源
@@ -311,7 +311,7 @@ public class MediaProjectionService extends Service {
             case NAL_I: {
                 byte[] newData = new byte[configFrameCase.length + data.length + 1];
                 Log.e("liu1017", String.format("configFrameCase = %s    data = %s    newData = %s", configFrameCase.length, data.length, newData.length));
-                newData[0] = (byte) TYPE_VIDEO;
+                newData[0] = TYPE_VIDEO;
                 System.arraycopy(configFrameCase, 0, newData, 1, configFrameCase.length);
                 System.arraycopy(data, 0, newData, (1 + configFrameCase.length), data.length);
                 if (socketType == SOCKE_TYPE_SERVICE) {
@@ -324,7 +324,7 @@ public class MediaProjectionService extends Service {
             //
             default: {
                 byte[] newData = new byte[data.length + 1];
-                newData[0] = (byte) TYPE_VIDEO;
+                newData[0] = TYPE_VIDEO;
                 System.arraycopy(data, 0, newData, 1, data.length);
                 if (socketType == SOCKE_TYPE_SERVICE) {
                     SocketUtils.getInstance().serverSend(newData);
